KFSUPGRADE-174: Flip KimEntityExternalIdType is encrypted flag
We dont store any sensitive data in our kim tables and
the data is stored in clear text.  This data is fed in through
the kim feed and is not changed through rice docs.  However
we have turned on the encryption required flag for the tax id
field.  This causes constant problems with KIM trying to decrypt
plain text data.